# backend/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/suplier/update/{code}": {
            "put": {
                "tags": [
                    "Suplier"
                ],
                "summary": "Update Suplier",
                "operationId": "430e883648c76325bcbefdcfbe3e2887",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "code For Suplier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "All Request Body For Suplier",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PT komputer Jaya"
                                    },
                                    "contact_person": {
                                        "type": "string",
                                        "example": "Budi Tabuti"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "082233712925"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "budi@gmail.com"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl Flamboyan no 4"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Values",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}